Completed CvsApi project
The project is a REST API that takes in a CSV file with data in the format of, ID;Name;Age;Email, and returns de data in a JSON format with an endpoint at api/data. It allows for the parameter query "limit" in order to specify how many rows should be returned. If the limit parameter is not specified, all rows will be returned.